LPD-84844 Move the service tracker to module, it does not have to stay in portal-kernel as the two message listeners in exportimport service module are the only users
