+++ modules/apps/export-import/export-import-service/bnd.bnd
@@ -1,3 +1,4 @@
 Bundle-Name: Liferay Export Import Service
 Bundle-SymbolicName: com.liferay.exportimport.service
 Bundle-Version: 8.0.144
+-dsannotations-options: inherit
