+++ modules/apps/export-import/export-import-service/src/main/java/com/liferay/exportimport/lifecycle/BaseExportImportLifecycleMessageListener.java
@@ -5,14 +5,22 @@
 
 package com.liferay.exportimport.lifecycle;
 
+import com.liferay.exportimport.kernel.lifecycle.EventAwareExportImportLifecycleListener;
 import com.liferay.exportimport.kernel.lifecycle.ExportImportLifecycleEvent;
 import com.liferay.exportimport.kernel.lifecycle.ExportImportLifecycleListener;
+import com.liferay.exportimport.kernel.lifecycle.ExportImportLifecycleListenerFactoryUtil;
+import com.liferay.exportimport.kernel.lifecycle.ProcessAwareExportImportLifecycleListener;
+import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerList;
+import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerListFactory;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.messaging.BaseMessageListener;
 import com.liferay.portal.kernel.messaging.Message;
 
-import java.util.Set;
+import org.osgi.framework.BundleContext;
+import org.osgi.framework.ServiceReference;
+import org.osgi.service.component.annotations.Deactivate;
+import org.osgi.util.tracker.ServiceTrackerCustomizer;
 
 /**
  * @author Daniel Kocsis
@@ -20,17 +28,26 @@ import java.util.Set;
 public abstract class BaseExportImportLifecycleMessageListener
 	extends BaseMessageListener {
 
+	protected void activate(BundleContext bundleContext, boolean parallel) {
+		_serviceTrackerList = ServiceTrackerListFactory.open(
+			bundleContext, ExportImportLifecycleListener.class, null,
+			new ExportImportLifecycleListenerServiceTrackerCustomizer(
+				bundleContext, parallel));
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		_serviceTrackerList.close();
+	}
+
 	@Override
 	protected void doReceive(Message message) throws Exception {
-		Set<ExportImportLifecycleListener> exportImportLifecycleListeners =
-			getExportImportLifecycleListeners(message);
-
 		ExportImportLifecycleEvent exportImportLifecycleEvent =
 			(ExportImportLifecycleEvent)message.get(
 				"exportImportLifecycleEvent");
 
 		for (ExportImportLifecycleListener exportImportLifecycleListener :
-				exportImportLifecycleListeners) {
+				_serviceTrackerList) {
 
 			try {
 				exportImportLifecycleListener.onExportImportLifecycleEvent(
@@ -47,10 +64,73 @@ public abstract class BaseExportImportLifecycleMessageListener
 		}
 	}
 
-	protected abstract Set<ExportImportLifecycleListener>
-		getExportImportLifecycleListeners(Message message);
-
 	private static final Log _log = LogFactoryUtil.getLog(
 		BaseExportImportLifecycleMessageListener.class);
+
+	private ServiceTrackerList<ExportImportLifecycleListener>
+		_serviceTrackerList;
+
+	private static class ExportImportLifecycleListenerServiceTrackerCustomizer
+		implements ServiceTrackerCustomizer
+			<ExportImportLifecycleListener, ExportImportLifecycleListener> {
+
+		@Override
+		public ExportImportLifecycleListener addingService(
+			ServiceReference<ExportImportLifecycleListener> serviceReference) {
+
+			ExportImportLifecycleListener exportImportLifecycleListener =
+				_bundleContext.getService(serviceReference);
+
+			if (exportImportLifecycleListener instanceof
+					ProcessAwareExportImportLifecycleListener) {
+
+				exportImportLifecycleListener =
+					ExportImportLifecycleListenerFactoryUtil.create(
+						(ProcessAwareExportImportLifecycleListener)
+							exportImportLifecycleListener);
+			}
+			else if (exportImportLifecycleListener instanceof
+						EventAwareExportImportLifecycleListener) {
+
+				exportImportLifecycleListener =
+					ExportImportLifecycleListenerFactoryUtil.create(
+						(EventAwareExportImportLifecycleListener)
+							exportImportLifecycleListener);
+			}
+
+			if (exportImportLifecycleListener.isParallel() != _parallel) {
+				_bundleContext.ungetService(serviceReference);
+
+				return null;
+			}
+
+			return exportImportLifecycleListener;
+		}
+
+		@Override
+		public void modifiedService(
+			ServiceReference<ExportImportLifecycleListener> serviceReference,
+			ExportImportLifecycleListener exportImportLifecycleListener) {
+		}
+
+		@Override
+		public void removedService(
+			ServiceReference<ExportImportLifecycleListener> serviceReference,
+			ExportImportLifecycleListener exportImportLifecycleListener) {
+
+			_bundleContext.ungetService(serviceReference);
+		}
+
+		private ExportImportLifecycleListenerServiceTrackerCustomizer(
+			BundleContext bundleContext, boolean parallel) {
+
+			_bundleContext = bundleContext;
+			_parallel = parallel;
+		}
+
+		private final BundleContext _bundleContext;
+		private final boolean _parallel;
+
+	}
 
 }
